Alignment to GAF; not tested
